// packages/frontend/store/modals/action-types.ts
export enum VerifyActionTypes {
  SHOW_VERIFY_PHONE_MODAL = 'verify-phone/show-modal',
  HIDE_VERIFY_PHONE_MODAL = 'verify-phone/hide-modal',
  NEXT_VERIFY_PHONE_MODAL = 'verify-phone/next-modal',
  PREVIOUS_VERIFY_PHONE_MODAL = 'verify-phone/previous-modal',
  SHOW_VERIFY_EMAIL_MODAL = 'verify-email/show-modal',
  HIDE_VERIFY_EMAIL_MODAL = 'verify-email/hide-modal',
  NEXT_VERIFY_EMAIL_MODAL = 'verify-email/next-modal',
  PREVIOUS_VERIFY_EMAIL_MODAL = 'verify-email/previous-modal',
  SHOW_MAKE_POST_MODAL = 'make-post/show-modal',
  HIDE_MAKE_POST_MODAL = 'make-post/hide-modal',
}